old tournament bracket file removed, replaced with new tournament bracket file
